fix: bugs to build, refactor: pages
