+++ src/pages/adminPageCandidateList/index.tsx
@@ -5,6 +5,7 @@ import add from "../../assets/img/addIcons.svg";
 import sub from "../../assets/img/subIcons.svg";
 import { v4 as uuidv4 } from 'uuid';
 
+
 export const AdminPageCandidateList = () => {
 
     const { candidateList, candidates, findJobsByCompanyId, specificJobs } = useContext(AdminContext)
@@ -23,7 +24,7 @@ export const AdminPageCandidateList = () => {
 
     const candidatureList = candidates.map((candidate) => {
 
-        const job = specificJobs.find((job) => job.id === candidate.jobId);
+        const job = specificJobs.find((job) => job.data.id === candidate.data.jobId);
 
         return {
             candidate,
@@ -44,7 +45,7 @@ export const AdminPageCandidateList = () => {
                 <ul>
                     {candidates ? candidatureList.map((e) => <li key={cryptoUUID()}>
                         <div className="candidateNameDiv" onClick={() => handdleShowJob()}>
-                            <h2>{e.candidate.name} - {e.job.position}</h2>
+                            <h2>{e.candidate.data.name} - {e.job?.data.position}</h2>
                             <div>
                                 {showJob ? (
                                     <img src={sub} className="subIcon" alt="sub icons" />
@@ -55,8 +56,8 @@ export const AdminPageCandidateList = () => {
                         </div>
                         <div className={showJob ? "descriptionCandidateDiv" : "descriptionCandidateDivNone"}>
                             <h3 className="titleCandidatureDetails">Detalhes da candidatura:</h3>
-                            <p>{e.job.description}</p>
-                            <p>E-mail: <span className="candidateEmail">{e.candidate.email}</span></p>
+                            <p>{e.job?.data.description}</p>
+                            <p>E-mail: <span className="candidateEmail">{e.candidate.data.email}</span></p>
                         </div>
                     </li>) : <div><h1>Nenhum candidato foi cadastrado até o momento</h1></div>}
                 </ul>
